refactor: reuse code shared between pupil pages
