+++ src/pages/pupils/l/[redirectFrom]/lessons/[lessonSlug]/index.tsx
@@ -12,8 +12,7 @@ import {
   shouldSkipInitialBuild,
 } from "@/node-lib/isr";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
-import { getCaptionsFromFile } from "@/utils/handleTranscript";
-import getDownloadResourcesExistence from "@/components/SharedComponents/helpers/downloadAndShareHelpers/getDownloadResourcesExistence";
+import { requestLessonResources } from "@/components/PupilComponents/pupilUtils/requestLessonResources";
 import {
   PupilExperienceView,
   PupilExperienceViewProps,
@@ -100,27 +99,8 @@ export const getStaticProps: GetStaticProps<
         };
       }
 
-      // For new content we need to fetch the captions file from gCloud and parse the result to generate
-      // the transcript sentences.
-      const resolveTranscriptSentences = (() => {
-        if (curriculumData.videoTitle && !curriculumData.isLegacy) {
-          return getCaptionsFromFile(`${curriculumData.videoTitle}.vtt`);
-        }
-
-        return curriculumData.transcriptSentences;
-      })();
-
-      // Resolve the requests for the transcript and worksheet existence in parallel
-      const [transcriptSentences, downloadExistence] = await Promise.all([
-        resolveTranscriptSentences,
-        curriculumData.isLegacy
-          ? { resources: [] }
-          : getDownloadResourcesExistence(
-              lessonSlug,
-              "worksheet-pdf",
-              curriculumData.isLegacy,
-            ),
-      ]);
+      const { transcriptSentences, hasWorksheet } =
+        await requestLessonResources({ curriculumData });
 
       const results: GetStaticPropsResult<PupilExperienceViewProps> = {
         props: {
@@ -128,9 +108,7 @@ export const getStaticProps: GetStaticProps<
             ...curriculumData,
             transcriptSentences: transcriptSentences ?? [],
           },
-          hasWorksheet: downloadExistence.resources.some(
-            ([type, result]) => type === "worksheet-pdf" && result.exists,
-          ),
+          hasWorksheet,
           backUrl: redirectUrl,
         },
       };
