+++ src/pages/pupils/lessons/[lessonSlug]/index.tsx
@@ -11,12 +11,11 @@ import {
   shouldSkipInitialBuild,
 } from "@/node-lib/isr";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
-import { getCaptionsFromFile } from "@/utils/handleTranscript";
-import getDownloadResourcesExistence from "@/components/SharedComponents/helpers/downloadAndShareHelpers/getDownloadResourcesExistence";
 import {
   PupilExperienceView,
   PupilExperienceViewProps,
 } from "@/components/PupilViews/PupilExperience";
+import { requestLessonResources } from "@/components/PupilComponents/pupilUtils/requestLessonResources";
 
 /**
  * Test URLs:
@@ -87,27 +86,8 @@ export const getStaticProps: GetStaticProps<
         };
       }
 
-      // For new content we need to fetch the captions file from gCloud and parse the result to generate
-      // the transcript sentences.
-      const resolveTranscriptSentences = (() => {
-        if (curriculumData.videoTitle && !curriculumData.isLegacy) {
-          return getCaptionsFromFile(`${curriculumData.videoTitle}.vtt`);
-        }
-
-        return curriculumData.transcriptSentences;
-      })();
-
-      // Resolve the requests for the transcript and worksheet existence in parallel
-      const [transcriptSentences, downloadExistence] = await Promise.all([
-        resolveTranscriptSentences,
-        curriculumData.isLegacy
-          ? { resources: [] }
-          : getDownloadResourcesExistence(
-              lessonSlug,
-              "worksheet-pdf",
-              curriculumData.isLegacy,
-            ),
-      ]);
+      const { transcriptSentences, hasWorksheet } =
+        await requestLessonResources({ curriculumData });
 
       const results: GetStaticPropsResult<PupilExperienceViewProps> = {
         props: {
@@ -115,9 +95,7 @@ export const getStaticProps: GetStaticProps<
             ...curriculumData,
             transcriptSentences: transcriptSentences ?? [],
           },
-          hasWorksheet: downloadExistence.resources.some(
-            ([type, result]) => type === "worksheet-pdf" && result.exists,
-          ),
+          hasWorksheet,
         },
       };
